requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
boto==2.49.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
DateTime==4.3
dnspython==2.1.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
MarkupSafe==1.1.1
nltk==3.5
pymongo==3.11.2
python-dotenv==0.15.0
pytz==2020.5
regex==2020.11.13
requests==2.25.1
six==1.15.0
soupsieve==2.1
textblob==0.15.3
tqdm==4.55.1
tzlocal==2.1
urllib3==1.26.2
Werkzeug==1.0.1
zope.interface==5.2.0
